migrated some old docs
