+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "horseless",
-  "version": "0.0.24",
+  "version": "0.1.0",
   "description": "a framework?",
   "main": "dist/horseless.cjs.js",
   "module": "dist/horseless.min.esm.js",
